requirements.txt
termcolor
tensorflow==2.12.1
scikit-learn
numpy
pillow
matplotlib
pyqt5
